+++ internal/goscrobble/album.go
@@ -3,17 +3,16 @@ package goscrobble
 import (
 	"database/sql"
 	"errors"
-	"fmt"
 	"log"
 )
 
 type Album struct {
-	Uuid          string         `json:"uuid"`
-	Name          string         `json:"name"`
-	Desc          sql.NullString `json:"desc"`
-	Img           sql.NullString `json:"img"`
-	MusicBrainzID string         `json:"mbid"`
-	SpotifyID     string         `json:"spotify_id"`
+	UUID          string `json:"uuid"`
+	Name          string `json:"name"`
+	Desc          string `json:"desc"`
+	Img           string `json:"img"`
+	MusicBrainzID string `json:"mbid"`
+	SpotifyID     string `json:"spotify_id"`
 }
 
 // insertAlbum - This will return if it exists or create it based on MBID > Name
@@ -22,32 +21,25 @@ func insertAlbum(name string, mbid string, spotifyId string, artists []string, t
 
 	// Try locate our album
 	if mbid != "" {
-		album = fetchAlbum("mbid", mbid, tx)
+		album = getAlbumByCol("mbid", mbid, tx)
 	} else if spotifyId != "" {
-		album = fetchAlbum("spotify_id", spotifyId, tx)
+		album = getAlbumByCol("spotify_id", spotifyId, tx)
 	}
 
 	// If it didn't match above, lookup by name
-	if album.Uuid == "" {
-		album = fetchAlbum("name", name, tx)
+	if album.UUID == "" {
+		album = getAlbumByCol("name", name, tx)
 	}
 
 	// If we can't find it. Lets add it!
-	if album.Uuid == "" {
-		err := insertNewAlbum(name, mbid, spotifyId, tx)
+	if album.UUID == "" {
+		err := insertNewAlbum(&album, name, mbid, spotifyId, tx)
 		if err != nil {
 			return album, errors.New("Failed to insert album")
 		}
 
-		// Fetch the recently inserted album to get the UUID
-		if mbid != "" {
-			album = fetchAlbum("mbid", mbid, tx)
-		} else if spotifyId != "" {
-			album = fetchAlbum("spotify_id", spotifyId, tx)
-		}
-
-		if album.Uuid == "" {
-			album = fetchAlbum("name", name, tx)
+		if album.UUID == "" {
+			return album, errors.New("Failed to fetch album")
 		}
 
 		// Try linkem up
@@ -57,10 +49,7 @@ func insertAlbum(name string, mbid string, spotifyId string, artists []string, t
 		}
 	}
 
-	if album.Uuid == "" {
-		return album, errors.New("Unable to fetch album!")
-	}
-
+	// Updates these values if we match earlier!
 	if album.MusicBrainzID != mbid {
 		album.MusicBrainzID = mbid
 		album.updateAlbum("mbid", mbid, tx)
@@ -74,11 +63,11 @@ func insertAlbum(name string, mbid string, spotifyId string, artists []string, t
 	return album, nil
 }
 
-func fetchAlbum(col string, val string, tx *sql.Tx) Album {
+func getAlbumByCol(col string, val string, tx *sql.Tx) Album {
 	var album Album
 	err := tx.QueryRow(
-		"SELECT BIN_TO_UUID(`uuid`, true), `name`, `desc`, `img`, `mbid` FROM `albums` WHERE `"+col+"` = ?",
-		val).Scan(&album.Uuid, &album.Name, &album.Desc, &album.Img, &album.MusicBrainzID)
+		"SELECT BIN_TO_UUID(`uuid`, true), `name`, IFNULL(`desc`, ''), IFNULL(`img`,''), `mbid`, `spotify_id` FROM `albums` WHERE `"+col+"` = ?",
+		val).Scan(&album.UUID, &album.Name, &album.Desc, &album.Img, &album.MusicBrainzID, &album.SpotifyID)
 
 	if err != nil {
 		if err != sql.ErrNoRows {
@@ -89,9 +78,14 @@ func fetchAlbum(col string, val string, tx *sql.Tx) Album {
 	return album
 }
 
-func insertNewAlbum(name string, mbid string, spotifyId string, tx *sql.Tx) error {
+func insertNewAlbum(album *Album, name string, mbid string, spotifyId string, tx *sql.Tx) error {
+	album.UUID = newUUID()
+	album.Name = name
+	album.MusicBrainzID = mbid
+	album.SpotifyID = spotifyId
+
 	_, err := tx.Exec("INSERT INTO `albums` (`uuid`, `name`, `mbid`, `spotify_id`) "+
-		"VALUES (UUID_TO_BIN(UUID(), true),?,?,?)", name, mbid, spotifyId)
+		"VALUES (UUID_TO_BIN(?, true),?,?,?)", album.UUID, album.Name, album.MusicBrainzID, album.SpotifyID)
 
 	return err
 }
@@ -100,9 +94,8 @@ func (album *Album) linkAlbumToArtists(artists []string, tx *sql.Tx) error {
 	var err error
 	for _, artist := range artists {
 		_, err = tx.Exec("INSERT INTO `album_artist` (`album`, `artist`) "+
-			"VALUES (UUID_TO_BIN(?, true), UUID_TO_BIN(?, true))", album.Uuid, artist)
+			"VALUES (UUID_TO_BIN(?, true), UUID_TO_BIN(?, true))", album.UUID, artist)
 		if err != nil {
-			fmt.Println(err)
 			return err
 		}
 	}
@@ -111,7 +104,19 @@ func (album *Album) linkAlbumToArtists(artists []string, tx *sql.Tx) error {
 }
 
 func (album *Album) updateAlbum(col string, val string, tx *sql.Tx) error {
-	_, err := tx.Exec("UPDATE `albums` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, album.Uuid)
+	_, err := tx.Exec("UPDATE `albums` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, album.UUID)
 
 	return err
+}
+
+func getAlbumByUUID(uuid string) (Album, error) {
+	var album Album
+	err := db.QueryRow("SELECT BIN_TO_UUID(`uuid`, true), `name`, IFNULL(`desc`,''), IFNULL(`img`,''), `mbid`, `spotify_id` FROM `albums` WHERE `uuid` = UUID_TO_BIN(?, true)",
+		uuid).Scan(&album.UUID, &album.Name, &album.Desc, &album.Img, &album.MusicBrainzID, &album.SpotifyID)
+
+	if err != nil {
+		return album, errors.New("Invalid UUID")
+	}
+
+	return album, nil
 }