+++ internal/goscrobble/track.go
@@ -3,19 +3,18 @@ package goscrobble
 import (
 	"database/sql"
 	"errors"
-	"fmt"
 	"log"
 	"strings"
 )
 
 type Track struct {
-	Uuid          string         `json:"uuid"`
-	Name          string         `json:"name"`
-	Length        int            `json:"length"`
-	Desc          sql.NullString `json:"desc"`
-	Img           sql.NullString `json:"img"`
-	MusicBrainzID string         `json:"mbid"`
-	SpotifyID     string         `json:"spotify_id"`
+	UUID          string `json:"uuid"`
+	Name          string `json:"name"`
+	Length        int    `json:"length"`
+	Desc          string `json:"desc"`
+	Img           string `json:"img"`
+	MusicBrainzID string `json:"mbid"`
+	SpotifyID     string `json:"spotify_id"`
 }
 
 // insertTrack - This will return if it exists or create it based on MBID > Name
@@ -24,35 +23,26 @@ func insertTrack(name string, legnth int, mbid string, spotifyId string, album s
 
 	// Try locate our track
 	if mbid != "" {
-		track = fetchTrack("mbid", mbid, tx)
-		fmt.Printf("Fetech mbid: %s", mbid)
+		track = getTrackByCol("mbid", mbid, tx)
 	} else if spotifyId != "" {
-		track = fetchTrack("spotify_id", spotifyId, tx)
-		fmt.Printf("Fetech spotify: %s", spotifyId)
+		track = getTrackByCol("spotify_id", spotifyId, tx)
 	}
 
 	// If it didn't match above, lookup by name
-	if track.Uuid == "" {
-		// TODO: add artist check here too
-		track = fetchTrackWithArtists(name, artists, album, tx)
+	if track.UUID == "" {
+		track = getTrackWithArtists(name, artists, album, tx)
 	}
 
 	// If we can't find it. Lets add it!
-	if track.Uuid == "" {
-		err := insertNewTrack(name, legnth, mbid, spotifyId, tx)
+	if track.UUID == "" {
+		err := insertNewTrack(&track, name, legnth, mbid, spotifyId, tx)
+
 		if err != nil {
 			return track, errors.New("Failed to insert track")
 		}
 
-		// Fetch the recently inserted track to get the UUID
-		if mbid != "" {
-			track = fetchTrack("mbid", mbid, tx)
-		} else if spotifyId != "" {
-			track = fetchTrack("spotify_id", spotifyId, tx)
-		}
-
-		if track.Uuid == "" {
-			track = fetchTrackWithArtists(name, artists, album, tx)
+		if track.UUID == "" {
+			return track, errors.New("Unable to fetch track!")
 		}
 
 		err = track.linkTrack(album, artists, tx)
@@ -61,10 +51,7 @@ func insertTrack(name string, legnth int, mbid string, spotifyId string, album s
 		}
 	}
 
-	if track.Uuid == "" {
-		return track, errors.New("Unable to fetch track!")
-	}
-
+	// Updates these values if we match earlier!
 	if track.MusicBrainzID != mbid {
 		track.MusicBrainzID = mbid
 		track.updateTrack("mbid", mbid, tx)
@@ -78,11 +65,11 @@ func insertTrack(name string, legnth int, mbid string, spotifyId string, album s
 	return track, nil
 }
 
-func fetchTrack(col string, val string, tx *sql.Tx) Track {
+func getTrackByCol(col string, val string, tx *sql.Tx) Track {
 	var track Track
 	err := tx.QueryRow(
-		"SELECT BIN_TO_UUID(`uuid`, true), `name`, `desc`, `img`, `mbid` FROM `tracks` WHERE `"+col+"` = ? LIMIT 1",
-		val).Scan(&track.Uuid, &track.Name, &track.Desc, &track.Img, &track.MusicBrainzID)
+		"SELECT BIN_TO_UUID(`uuid`, true), `name`, IFNULL(`desc`,''), IFNULL(`img`,''), `mbid` FROM `tracks` WHERE `"+col+"` = ? LIMIT 1",
+		val).Scan(&track.UUID, &track.Name, &track.Desc, &track.Img, &track.MusicBrainzID)
 
 	if err != nil {
 		if err != sql.ErrNoRows {
@@ -93,16 +80,16 @@ func fetchTrack(col string, val string, tx *sql.Tx) Track {
 	return track
 }
 
-func fetchTrackWithArtists(name string, artists []string, album string, tx *sql.Tx) Track {
+func getTrackWithArtists(name string, artists []string, album string, tx *sql.Tx) Track {
 	var track Track
 	artistString := strings.Join(artists, "','")
 	err := tx.QueryRow(
-		"SELECT BIN_TO_UUID(`uuid`, true), `name`, `desc`, `img`, `mbid` FROM `tracks` "+
+		"SELECT BIN_TO_UUID(`uuid`, true), `name`, IFNULL(`desc`,''), IFNULL(`img`,''), `mbid` FROM `tracks` "+
 			"LEFT JOIN `track_artist` ON `tracks`.`uuid` = `track_artist`.`track` "+
 			"LEFT JOIN `track_album` ON `tracks`.`uuid` = `track_album`.`track` "+
 			"WHERE `name` = ? AND BIN_TO_UUID(`track_artist`.`artist`, true) IN ('"+artistString+"') "+
 			"AND BIN_TO_UUID(`track_album`.`album`,true) = ? LIMIT 1",
-		name, album).Scan(&track.Uuid, &track.Name, &track.Desc, &track.Img, &track.MusicBrainzID)
+		name, album).Scan(&track.UUID, &track.Name, &track.Desc, &track.Img, &track.MusicBrainzID)
 
 	if err != nil {
 		if err != sql.ErrNoRows {
@@ -113,9 +100,15 @@ func fetchTrackWithArtists(name string, artists []string, album string, tx *sql.
 	return track
 }
 
-func insertNewTrack(name string, length int, mbid string, spotifyId string, tx *sql.Tx) error {
+func insertNewTrack(track *Track, name string, length int, mbid string, spotifyId string, tx *sql.Tx) error {
+	track.UUID = newUUID()
+	track.Name = name
+	track.Length = length
+	track.MusicBrainzID = mbid
+	track.SpotifyID = spotifyId
+
 	_, err := tx.Exec("INSERT INTO `tracks` (`uuid`, `name`, `length`, `mbid`, `spotify_id`) "+
-		"VALUES (UUID_TO_BIN(UUID(), true),?,?,?,?)", name, length, mbid, spotifyId)
+		"VALUES (UUID_TO_BIN(?, true),?,?,?,?)", track.UUID, track.Name, track.Length, track.MusicBrainzID, track.SpotifyID)
 
 	return err
 }
@@ -126,24 +119,26 @@ func (track *Track) linkTrack(album string, artists []string, tx *sql.Tx) error 
 		return err
 	}
 	err = track.linkTrackToArtists(artists, tx)
+
 	if err != nil {
 		return err
 	}
+
 	return nil
 }
 
-func (track Track) linkTrackToAlbum(albumUuid string, tx *sql.Tx) error {
+func (track *Track) linkTrackToAlbum(albumUuid string, tx *sql.Tx) error {
 	_, err := tx.Exec("INSERT INTO `track_album` (`track`, `album`) "+
-		"VALUES (UUID_TO_BIN(?, true), UUID_TO_BIN(?, true))", track.Uuid, albumUuid)
+		"VALUES (UUID_TO_BIN(?, true), UUID_TO_BIN(?, true))", track.UUID, albumUuid)
 
 	return err
 }
 
-func (track Track) linkTrackToArtists(artists []string, tx *sql.Tx) error {
+func (track *Track) linkTrackToArtists(artists []string, tx *sql.Tx) error {
 	var err error
 	for _, artist := range artists {
 		_, err = tx.Exec("INSERT INTO `track_artist` (`track`, `artist`) "+
-			"VALUES (UUID_TO_BIN(?, true),UUID_TO_BIN(?, true))", track.Uuid, artist)
+			"VALUES (UUID_TO_BIN(?, true),UUID_TO_BIN(?, true))", track.UUID, artist)
 		if err != nil {
 			return err
 		}
@@ -153,7 +148,19 @@ func (track Track) linkTrackToArtists(artists []string, tx *sql.Tx) error {
 }
 
 func (track *Track) updateTrack(col string, val string, tx *sql.Tx) error {
-	_, err := tx.Exec("UPDATE `tracks` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, track.Uuid)
+	_, err := tx.Exec("UPDATE `tracks` SET `"+col+"` = ? WHERE `uuid` = UUID_TO_BIN(?,true)", val, track.UUID)
 
 	return err
+}
+
+func getTrackByUUID(uuid string) (Track, error) {
+	var track Track
+	err := db.QueryRow("SELECT BIN_TO_UUID(`uuid`, true), `name`, IFNULL(`desc`,''), IFNULL(`img`,''), `length`, `mbid`, `spotify_id` FROM `tracks` WHERE `uuid` = UUID_TO_BIN(?, true)",
+		uuid).Scan(&track.UUID, &track.Name, &track.Desc, &track.Img, &track.Length, &track.MusicBrainzID, &track.SpotifyID)
+
+	if err != nil {
+		return track, errors.New("Invalid UUID")
+	}
+
+	return track, nil
 }